requirements.txt
sounddevice==0.4.0
pyzmq==19.0.2
SoundFile==0.10.3
PyYAML==5.3.1
